refactor(profile/layouts): move checking search params logic to `StatsOptions` component
