+++ app/profile/[id]/@top/layout.tsx
@@ -1,22 +1,9 @@
 'use client'
 
-import { useSearchParams } from 'next/navigation'
 import type { ReactNode } from 'react'
 
 import { StatsOptions } from '@app/profile/components/common'
-import {
-  STATS_MEASUREMENT,
-  STATS_PROVIDER,
-  TIME_RANGE,
-  VIEW,
-} from '@app/profile/constants'
 import type { ProfileLayoutBaseProps } from '@app/profile/types'
-import {
-  validateStatsMeasurement,
-  validateStatsProvider,
-  validateTimeRange,
-  validateView,
-} from '@app/profile/utils/validators'
 
 namespace ProfileTopGenresSubLayout {
   export interface Props extends ProfileLayoutBaseProps {
@@ -33,25 +20,9 @@ function ProfileTopGenresSubLayout({
   tracks,
   albums,
 }: ProfileTopGenresSubLayout.Props) {
-  const searchParams = useSearchParams()
-  const statsProvider = validateStatsProvider(searchParams.get(STATS_PROVIDER))
-  const timeRange = validateTimeRange(
-    searchParams.get(TIME_RANGE),
-    statsProvider
-  )
-  const statsMeasurement = validateStatsMeasurement(
-    searchParams.get(STATS_MEASUREMENT)
-  )
-  const view = validateView(searchParams.get(VIEW))
-
   return (
     <>
-      <StatsOptions
-        statsProvider={statsProvider}
-        statsMeasurement={statsMeasurement}
-        view={view}
-        timeRange={timeRange}
-      />
+      <StatsOptions />
 
       {genres}
       {artists}
